requirements.txt
discord.py>=2.3.0
python-dotenv>=0.21.0
openai>=0.28.0
aiohttp>=3.8.1
requests>=2.28.0
beautifulsoup4>=4.12.0
lxml>=4.9.2
flask>=2.2.0
Markdown>=3.4.0
pandas>=2.0.0
openpyxl>=3.0.10
xlrd>=2.0.1
numpy>=1.24.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
